+++ cmd/sites_meta.go
@@ -15,7 +15,7 @@ func setSiteCommand() *cobra.Command {
 		if err := system.RequireRoot(); err != nil {
 			return permissionError{err}
 		}
-		cfg, err := config.Load("")
+		cfg, err := loadManagedConfig()
 		if err != nil {
 			return err
 		}
@@ -60,7 +60,7 @@ func archiveSetCommand(name string, archived bool) *cobra.Command {
 		if err := system.RequireRoot(); err != nil {
 			return permissionError{err}
 		}
-		cfg, err := config.Load("")
+		cfg, err := loadManagedConfig()
 		if err != nil {
 			return err
 		}